scripts/insert_table/requirements.txt
pandas
requests
boto3
psycopg2-binary
sqlalchemy
